requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
Jinja2==3.0.3
itsdangerous==2.0.1
MarkupSafe==2.0.1
requests
PyYAML